Write a SysRes value as an unsigned number. Previously, whether the bitpattern was interpreted as a signed or unsigned value depended on the platform.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@15531

+++ coregrind/m_syswrap/syswrap-main.c
@@ -2612,12 +2612,12 @@ VG_(fixup_guest_state_after_syscall_interrupted)( ThreadId tid,
    if (VG_(clo_trace_signals))
       VG_(message)( Vg_DebugMsg,
                     "interrupted_syscall: tid=%u, ip=%#lx, "
-                    "restart=%s, sres.isErr=%s, sres.val=%lld\n",
+                    "restart=%s, sres.isErr=%s, sres.val=%lu\n",
                     tid,
                     ip,
                     restart ? "True" : "False",
                     sr_isError(sres) ? "True" : "False",
-                    (Long)(sr_isError(sres) ? sr_Err(sres) : sr_Res(sres)) );
+                    sr_isError(sres) ? sr_Err(sres) : sr_Res(sres));
 
    vg_assert(VG_(is_valid_tid)(tid));
    vg_assert(tid >= 1 && tid < VG_N_THREADS);
